AppleCal is basically CalDav
